(untested) fix for the problem you described on the forum.  details below:
Hi, thanks for mentioning this.  In the 1.2 cleanup the 'data' layer of
packets 'disappeared' - moved one level higher.  So I removed that from
your code.  I bet this fixes things.

Also, in general in python it is good practice to use the 'modern'
exception catching syntax.  i.e. always get the exception as an object
and then (optionally) print it out (or even use the methods on ex to
get a stack trace).  I'd recommend you do this any place where you have
try/except blocks.  This would make the problem much easier to debug.